Cleanup
Fixed the dialog naming issue
Added an About dialog (HTML launch would not work)
Added a copy to clipboard upon generation and a manual button
