+++ StrongPassGenerator/GeneratePasswordForm.resx
