+++ StrongPassGenerator/Program.cs
@@ -19,7 +19,7 @@ namespace StrongPassGenerator
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new Form1());
+            Application.Run(new GeneratePasswordForm());
         }
     }
 }
